Update App.js
Implemented a basic API request from a random online database using axios so as to I do not have to explicitly transform the fetched data into JSON format.
Furthermore, I have implemented two functions(getData() and delCard()) that commit changes to the react state (data). setData has to be used because it's a functional component.